requirements.txt
streamlit
requests
beautifulsoup4
azure-ai-textanalytics
